refacto github bot with plugins
